:sparkles::construction: [WIP] Implement basic Profiles structure.
