+++ src/Command/Initializer/Profiles/blank/Init.php
@@ -1,0 +1,34 @@
+<?php
+
+/**
+ * Slim Framework (https://slimframework.com)
+ *
+ * @license https://github.com/slimphp/Slim-Console/blob/0.x/LICENSE.md (MIT License)
+ */
+
+declare(strict_types=1);
+
+namespace Slim\Console\Command\Initializer\Profiles\blank;
+
+use Slim\Console\Command\Initializer\Profiles\InitProfileInterface;
+use Symfony\Component\Console\Input\InputInterface;
+use Symfony\Component\Console\Output\OutputInterface;
+
+/**
+ * Init class implementation for profile Blank.
+ *
+ * @package Slim\Console\Command\Initializer\Profiles\blank
+ * @author Temuri Takalandze <me@abgeo.dev>
+ */
+class Init implements InitProfileInterface
+{
+    /**
+     * {@inheritDoc}
+     */
+    public function run(InputInterface $input, OutputInterface $output, bool $useDefaultSetup = false): int
+    {
+        $output->writeln('Work In Progress!');
+
+        return 0;
+    }
+}
