+++ src/Command/Initializer/Profiles/blank/PROFILE.md
@@ -1,0 +1,7 @@
+# Profile Blank
+
+Profile doc goes here...
+
+## Authors
+
+- [**Temuri Takalandze**](https://abgeo.dev)
